Item 8348 : Link QC and experimental data - display experiment run range region on QC plots (#331)

+++ src/org/labkey/targetedms/TargetedMSSchema.java
@@ -775,7 +775,7 @@ public class TargetedMSSchema extends UserSchema
                             @Override
                             public String renderURL(RenderContext ctx)
                             {
-                                Integer runId = ctx.get(FieldKey.fromParts("File", "Id"), Integer.class);
+                                Long runId = ctx.get(FieldKey.fromParts("File", "Id"), Long.class);
                                 if (runId == null)
                                     return null;
 
@@ -1324,7 +1324,7 @@ public class TargetedMSSchema extends UserSchema
         {
             try
             {
-                TargetedMSRun run = TargetedMSManager.getRun(Integer.parseInt(name.substring(SAMPLE_FILE_RUN_PREFIX.length())));
+                TargetedMSRun run = TargetedMSManager.getRun(Long.parseLong(name.substring(SAMPLE_FILE_RUN_PREFIX.length())));
                 // If we can't find the run or it's not in the current container, just act like the table doesn't exist
                 if (run != null && run.getContainer().equals(getContainer()))
                 {
